TINYMCE-14527: move resizing behavior to alloy
