+++ modules/alloy/src/main/ts/ephox/alloy/behaviour/resizing/ResizingSchema.ts
@@ -1,0 +1,5 @@
+import { type FieldProcessor, FieldSchema } from '@ephox/boulder';
+
+export default [
+  FieldSchema.requiredFunction('resize')
+] as FieldProcessor[];
